Change band precomputation to save complete bands
Previously only the positions of new points were saved, not the actual
DEM ids for entire bands. This results in exponentially larger files.
However it also removes repeated and therefore redundant calculations
from the actual TVS calculations. Precomputation is now slower, but it
is only ever done once for the whole planet, whereas TVS calculations,
which must be done for every DEM, are now faster. The cost of larger
pre-computation files is of course relevant, but disk space is a far
more abundant and workable resource than CPU cycles.