Compress band data with RLE [occurences,pattern..]
This was somewhat of an unexpected boon. After experimenting with
running the code on some big Google GPU VMs I realised that the band
building was a significant bottleneck, so decided to save the entire
pre-built bands during precomputation. This of course created a huge
amount of disk data (but improved the bottleneck). Whilst searching
through the corpus of dedicated integer compression algorithms it
occured to me that the band data here might have regular patterns in the
deltas (differences between DEM IDs) and indeed it did! So this commit
represents the simplest implementation of Run Length Encoded compressed
band data. It appears to compress to about a factor of 7.

Although there is some trivial rebuilding costs in the kernel, they are
small and I offset by 2 things. Firstly reduction in data transfer time
from host to device. Secondly in global memory accesses on the device,
as often DEM IDs can be calculated by mere addition, rather than fetched
from memory.