Switch to describing bands with deltas
This is a huge breakthrough, in fact possibly one of the most
significant commits of my life. Previously each and every band
was calculated individually. This required N^2 CPU during
precomputation and similarly N^2 for space in both precompute and
compute itself. Switching to using band deltas means only 1 half of a
band ever needs to be computed - all other bands are generated from the
shape of that band. This changes N^2 to SQRT(N). The effect is most useful
in terms of space requirements, however it also dramatically speeds up
precomputation, whilst the impact on computation is merely in terms of
the time saved by having to transfer less data to the GPU device.

It seems so obvious in hindsight and I'm confused as to why I haven't
seen it anywhere in the literature. This could possibly be a
contribution to the state of the art.

Note that as the transfer of band data is no longer the main bottleneck,
ring sector data is now the biggest user of RAM and disk. So I added
batching to the GPU computations, purely to manage the data generated
by the GPU.