refactor: CLI args, logging, README
First Rust version that loads tiles and computes. Just on the CPU for
now.

Also fixed a bug that I think was to do with bands of sight not being
filled enought, and at certain angles the line of sight was jumping up
to 20 rows (on SRTM3 files) because there weren't enough points in the
band of sight to hit.